casts, test

+++ ww/base/wmutex.h
@@ -128,8 +128,8 @@ static inline void timespec_after(struct timespec *ts, unsigned int ms)
 {
     struct timeval tv;
     getTimeOfDay(&tv, NULL);
-    ts->tv_sec  = tv.tv_sec + ms / 1000;
-    ts->tv_nsec = tv.tv_usec * 1000 + ms % 1000 * 1000000;
+    ts->tv_sec  = tv.tv_sec + ((long)ms / 1000);
+    ts->tv_nsec = tv.tv_usec * 1000 + (long)ms % 1000 * 1000000;
     if (ts->tv_nsec >= 1000000000)
     {
         ts->tv_nsec -= 1000000000;
